FFTW solver for uniform and 1-rank simulations
The solver seems to work but has this "tol" parameter that switches
between spectral (?) and finite-difference (?) approach. It is not clear
which produces better result in what cases.

If tol is set to 0.0 (fully finite difference), it is clear that odd and
even cells are effectively solved independenlty, which can even be seen
in simulation flow plots. On the other hand, tol of 1.0 (fully spectral)
produces a smooth result, but according to the original comment (commit
4c53ef1), produced non-div-free result, or at least non-div-free
presumably with respect to the 5-point stencil.

Also, another name should be picked for the "tol" parameter, if it is
kept at all.